alacritty: Improve update-script

+++ bin/update-alacritty.sh
@@ -1,14 +1,22 @@
 #!/bin/bash -x
-cd $1
-HASH=$(git rev-parse HEAD)
+
+if [ ! -d "$HOME/.cache" ]; then
+    mkdir ~/.cache
+fi
+if [ ! -d "$HOME/.cache/alacrittybuild" ]; then
+    git clone https://github.com/alacritty/alacritty ~/.cache/alacrittybuild
+    HASH=null
+fi
+cd ~/.cache/alacrittybuild
+: "${HASH:=$(git rev-parse HEAD)}"
 git pull
 if [ $HASH == $(git rev-parse HEAD) ]; then
-    echo "Alacritty is already up-to-date."
     exit 0
 fi
 make app
 if [ $? -eq 0 ]; then
-    echo 'Replace /Applications/Alacritty.app'
-    rm -r /Applications/Alacritty.app
-    mv ./target/release/osx/Alacritty.app /Applications/Alacritty.app
+    if [ -d "/Applications/Alacritty.app" ]; then
+        rm -r /Applications/Alacritty.app
+    fi
+    cp -r ./target/release/osx/Alacritty.app /Applications/Alacritty.app
 fi
